Tier B Phase 2 (Slice 2): move iframe-sandbox-policy to @mcpjam/sdk
Relocate the framework-free SEP-1865 iframe sandbox-attribute builders
(DEFAULT_IFRAME_SANDBOX, buildOuterAllowAttribute, buildOuterSandboxAttribute,
resolveIframeSandboxPolicy) into @mcpjam/sdk/widget-runtime, leaving a
back-compat re-export shim at the old inspector path so consumers
(host-app-bridge, sandboxed-iframe, eval browser-harness host-page) are
unchanged. Pure relocation; behavior unchanged.

- sdk: new src/widget-runtime/iframe-sandbox-policy.ts + barrel re-exports;
  added @modelcontextprotocol/ext-apps@^1.7.2 as a dependency (the module
  references the McpUiResourcePermissions type; Slice 3's host-app-bridge will
  use ext-apps as a value too). Extended test:packaging to verify the new
  exports.
- NOTE (beyond a pure file-move, for review): switched the SDK tsconfig from
  module/moduleResolution NodeNext -> ESNext/bundler. NodeNext can't follow
  ext-apps's `export * from "./types"` re-export chain for type-only members
  (TS2305/TS2460); `bundler` resolves it cleanly and matches how the SDK is
  actually shipped (tsup/esbuild) and how every consumer (the whole inspector)
  resolves. Emitted JS is produced by tsup and is unaffected; only the SDK's own
  typecheck/DTS resolution strategy changed.

Verification: full root `typecheck` (sdk+cli+design-system+chat-ui+mcp) passes;
sdk build + typecheck + test:packaging pass; inspector typecheck:client +
renderer tier-b guard pass; 134 widget tests pass (host-app-bridge 26,
sandboxed-iframe 22, mcp-apps-renderer 86); eval harness bundle resolves the
moved module; source alias resolves it with sdk/dist removed.
