+++ sdk/src/widget-runtime/index.ts
@@ -16,3 +16,10 @@ export {
 } from "./tool-visibility.js";
 
 export { LoggingTransport } from "./logging-transport.js";
+
+export {
+  DEFAULT_IFRAME_SANDBOX,
+  buildOuterAllowAttribute,
+  buildOuterSandboxAttribute,
+  resolveIframeSandboxPolicy,
+} from "./iframe-sandbox-policy.js";
